refactor JIT code
