+++ src/core/c/list.c
@@ -246,3 +246,46 @@ mobj list_append2(mobj xs, mobj ys) {
         return hd;
     }
 }
+
+mobj assq_ref(mobj xs, mobj k) {
+    for (; !minim_nullp(xs); xs = minim_cdr(xs)) {
+        if (minim_eqp(minim_caar(xs), k))
+            return minim_car(xs);
+    }
+
+    return minim_false;
+}
+
+mobj assq_set(mobj xs, mobj k, mobj v) {
+    if (minim_nullp(xs)) {
+        return Mlist1(Mcons(k, v));
+    } else {
+        mobj hd, tl;
+
+        if (minim_eqp(minim_caar(xs), k))
+            return Mcons(Mcons(minim_caar(xs), v), minim_cdr(xs));
+
+        hd = tl = Mcons(minim_car(xs), minim_null);
+        for (xs = minim_cdr(xs); !minim_nullp(xs); xs = minim_cdr(xs)) {
+            if (minim_eqp(minim_caar(xs), k)) {
+                minim_cdr(tl) = Mcons(Mcons(minim_caar(xs), v), minim_cdr(xs));
+                return hd;
+            } else {
+                minim_cdr(tl) = Mcons(minim_car(xs), minim_null);
+                tl = minim_cdr(tl);
+            }
+        }
+
+        minim_cdr(tl) = Mcons(Mcons(k, v), minim_null);
+        return hd;
+    }
+        
+    mobj cell = assq_ref(xs, k);
+    if (minim_falsep(cell)) {
+        return Mcons(Mcons(k, v), xs);
+    } else {
+
+    }
+
+    return Mcons(Mcons(k, v), xs);
+}
